requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.7
python-dotenv==1.0.0
web3==6.11.1
requests==2.31.0
cryptography==41.0.7
